+++ src/Sulu/Bundle/CategoryBundle/Tests/Unit/Content/Types/CategoryListTest.php
@@ -15,7 +15,7 @@ use JMS\Serializer\SerializationContext;
 use JMS\Serializer\SerializerInterface;
 use Prophecy\Argument;
 use Sulu\Bundle\CategoryBundle\Category\CategoryManagerInterface;
-use Sulu\Bundle\CategoryBundle\Entity\Category;
+use Sulu\Bundle\CategoryBundle\Entity\CategoryInterface;
 use Sulu\Component\Content\Compat\Property;
 use Sulu\Component\Content\Compat\StructureInterface;
 
@@ -50,11 +50,8 @@ class CategoryListTest extends \PHPUnit_Framework_TestCase
 
     public function testGetContentData()
     {
-        $entity1 = $this->prophesize(\Sulu\Bundle\CategoryBundle\Entity\CategoryInterface::class);
-        $entity2 = $this->prophesize(\Sulu\Bundle\CategoryBundle\Entity\CategoryInterface::class);
-
-        $category1 = $this->prophesize(Category::class);
-        $category2 = $this->prophesize(Category::class);
+        $category1 = $this->prophesize(CategoryInterface::class);
+        $category2 = $this->prophesize(CategoryInterface::class);
 
         $this->categoryManager->findByIds([1, 2])->willReturn([$category1, $category2]);